Moved int to long to be aligned with the demo 2^64 statement

+++ caesar-c/caesar.h
@@ -6,9 +6,9 @@
 
 /* Ciphers the input string using the provided shift */
 
-char *caesar_encrypt(char *str, int shift);
+char *caesar_encrypt(char *str, long shift);
 
 /* Deciphers the input ciphered string using the provided shift */
 
-char *caesar_decrypt(char *str, int shift);
+char *caesar_decrypt(char *str, long shift);
 #endif
